Remove "omitempty" from ConnectionOptionsSMS.From
When you use the API, you'll see that when you change the UI to select the Radio button to change between the two options, the "from" and "messaging_service_sid" get set to null for whichever one is not selected. If you set a "messaging_service_sid" without setting "from" to null, it will keep the radio button selected on "From" and simply error if the connection is used if no From number is configured ^Cen though a "messaging_service_sid" is properly configured.

Fixes #131