Add CustomText APIs
Add support for the https://auth0.com/docs/api/management/v2#!/Prompts/get_custom_text_by_language and https://auth0.com/docs/api/management/v2#!/Prompts/put_custom_text_by_language endpoints in Auth0.

The Get<Prompt>CustomText and Set<Prompt>CustomText methods call the corresponding GET and PUT /api/v2/prompts/{prompt}/custom-text/{language} and endpoints, where <Prompt> is the Camel Case version of the {prompt} id. For example, SetLoginIdCustomText("en") calls /api/v2/login-id/custom_text/en endpoint with the english ("en") language.

The APIs provide unit tests for each Get<Prompt>CustomText and Set<Prompt>CustomText pair using the English language as follows:

- Read StartVal using the Get<Prompt>CustomText method
- Initializes a Cleanup callback to restore StartVal when the test exits.
- Initialize SetVal using a value for each data member in the struct
- Set StartVal using the Set<Prompt>CustomText method
- Read ReturnVal using the Read<Prompt>CustomText method
- Assert that ReturnVal and SetVal are equivalent
- Execute the Cleanup method

Reading StartVal can return an empty struct, which causes an error in the Cleanup method in Set<Prompt>CustomText, for example:

    prompt_test.go:1404: 400 Bad Request: Screen "signup-password" must be a non-empty object.

Because of this, the error is logged as a warning instead of causing the unit tests to fail. The unit tests should not be run against a production environment, as there is no guarantee that the initial value in Auth0 will be restored properly -

The implementation is based on the prompt definitions in https://auth0.com/docs/universal-login/new-experience/text-customization-new-universal-login. When a documented struct data member is not found in the implementation, the struct data member has been commented with the notation "//Documented but not supported".